+++ app/src/main/java/cc/dlabs/pesamind/core/network/analytics/BudgetVsActualResponse.kt
@@ -12,10 +12,10 @@ data class BudgetVsActualResponse(
 data class BudgetActualData(
     val period: String,
     @SerializedName("budget_total")
-    val budgetTotal: Long,
+    val budgetTotal: Double,
     @SerializedName("actual_total")
-    val actualTotal: Long,
-    val variance: Long,
+    val actualTotal: Double,
+    val variance: Double,
     @SerializedName("variance_percent")
     val variancePercent: Double,
     val status: String,
@@ -28,9 +28,9 @@ data class BudgetActualData(
 
 data class BudgetActualItem(
     val category: String,
-    val budget: Long,
-    val actual: Long,
-    val variance: Long,
+    val budget: Double,
+    val actual: Double,
+    val variance: Double,
     @SerializedName("variance_percent")
     val variancePercent: Double,
     val transactions: Int,
